pyproject.toml
[tool.poetry]
name = "locust-cache-benchmark"
version = "1.0.0"
description = "A tool to perform load testing of Redis and Valkey systems using Locust"
authors = ["S-mishina"]
readme = "README.md"
homepage = "https://github.com/S-mishina/locust-cache-benchmark"
repository = "https://github.com/S-mishina/locust-cache-benchmark"
packages = [{include = "cache_benchmark", from = "cache_benchmark/src"}]

[tool.poetry.scripts]
locust_cache_benchmark = "cache_benchmark.main:main"

[tool.poetry.dependencies]
python = "^3.10"
locust = "^2.37.3"
redis = "^6.1.0"
valkey = "^6.1.0"
gevent = "^25.0.0"
psutil = "^7.0.0"
greenlet = "^3.2.2"
geventhttpclient = "^2.3.4"
msgpack = "^1.1.0"
pyzmq = "^27.0.0"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-instrumentation-redis = "^0.50b0"
opentelemetry-exporter-otlp = "^1.29.0"
opentelemetry-proto = "^1.29.0"
opentelemetry-semantic-conventions = "^0.50b0"
certifi = "^2025.4.26"
charset-normalizer = "^3.4.2"
idna = "^3.10"
packaging = "^25.0"
typing_extensions = "^4.13.2"
pydantic = "^2.12.5"
pydantic-settings = "^2.13.0"
python-json-logger = "^3.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^7.0.0"
coverage = "^7.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["cache_benchmark/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
